Move non-redux request generator to utils
Refs #379
